Dashboard/requirements.txt
pandas
dash
plotly
influxdb-client
PyYAML
tensorflow
numpy
joblib
dash-bootstrap-components
datetime
scikit-learn
python-dotenv